Consolidate all DPoP tests into DPoPLoginTests
Move DPoP-related tests from RTRLoginTests, MultiUserLoginTests,
RefreshTokenMigrationTests, and LoginWithRestartTests into DPoPLoginTests,
which now owns all DPoP coverage end-to-end (basic login, RTR, multi-user,
migration, restart). This mirrors how RTRLoginTests owns all RTR scenarios.

Update README, spec, and plan to reflect the consolidated structure.
