+++ native/NativeSampleApps/AuthFlowTester/src/androidTest/java/com/salesforce/samples/authflowtester/MultiUserLoginTests.kt
@@ -37,7 +37,6 @@ import com.salesforce.samples.authflowtester.testUtility.KnownAppConfig.BEACON_J
 import com.salesforce.samples.authflowtester.testUtility.KnownAppConfig.BEACON_OPAQUE
 import com.salesforce.samples.authflowtester.testUtility.KnownAppConfig.CA_OPAQUE
 import com.salesforce.samples.authflowtester.testUtility.KnownAppConfig.ECA_JWT
-import com.salesforce.samples.authflowtester.testUtility.KnownAppConfig.ECA_JWT_DPOP
 import com.salesforce.samples.authflowtester.testUtility.KnownAppConfig.ECA_OPAQUE
 import com.salesforce.samples.authflowtester.testUtility.KnownLoginHostConfig
 import com.salesforce.samples.authflowtester.testUtility.KnownLoginHostConfig.ADVANCED_AUTH
@@ -212,32 +211,6 @@ class MultiUserLoginTests: AuthFlowTest() {
         // Switch back to other user
         switchToUserAndValidate(otherUser)
         app.validateOAuthValues(knownAppConfig = ECA_JWT, scopeSelection = EMPTY)
-    }
-
-    // Both users log in with a DPoP ECA; tokens are unique and revoke+refresh works per-user.
-    @Test
-    fun testECAJwtDPoP_MultiUser_UniqueTokens() {
-        // Initial user with DPoP
-        loginAndValidate(knownAppConfig = ECA_JWT_DPOP, useDPoP = true)
-        val (userAccessToken, userRefreshToken) = app.getTokens()
-
-        // Other user with DPoP
-        loginOtherUserAndValidate(knownAppConfig = ECA_JWT_DPOP, useDPoP = true)
-        val (otherUserAccessToken, otherUserRefreshToken) = app.getTokens()
-
-        // Tokens must be unique across users
-        assertNotEquals(userAccessToken, otherUserAccessToken)
-        assertNotEquals(userRefreshToken, otherUserRefreshToken)
-
-        // Switch back to initial user; revoke + refresh must work with DPoP nonce rotation
-        switchToUserAndValidate(user)
-        app.validateOAuthValues(knownAppConfig = ECA_JWT_DPOP, scopeSelection = EMPTY)
-        assertRevokeAndRefreshWorks(isRtr = false, isDpop = true)
-
-        // Switch to other user; revoke + refresh must work independently with its own nonce
-        switchToUserAndValidate(otherUser)
-        app.validateOAuthValues(knownAppConfig = ECA_JWT_DPOP, scopeSelection = EMPTY)
-        assertRevokeAndRefreshWorks(isRtr = false, isDpop = true)
     }
 
     // Test MultiUser Token Migration.  This test also demonstrates the app restart validation
